Move charset loading to fix issue with static fields
